+++ react/components/WordpressSearchResultBlock.tsx
@@ -27,6 +27,7 @@ const WordpressSearchResultBlock: StorefrontFunctionComponent<WPSearchResultBloc
   subcategoryUrls,
   absoluteLinks,
   numberOfPosts,
+  mediaSize,
   customDomain,
   customDomainSlug,
 }) => {
@@ -69,9 +70,8 @@ const WordpressSearchResultBlock: StorefrontFunctionComponent<WPSearchResultBloc
                   excerpt={post.excerpt.rendered}
                   categories={post.categories}
                   subcategoryUrls={subcategoryUrls}
-                  image={post.featured_media?.source_url ?? ''}
-                  altText={post.featured_media?.alt_text ?? ''}
-                  mediaType={post.featured_media?.media_type ?? ''}
+                  featuredMedia={post.featured_media}
+                  mediaSize={mediaSize}
                   showCategory={showCategories}
                   showDate={showDates}
                   showAuthor={showAuthors}
@@ -116,6 +116,7 @@ interface WPSearchResultBlockProps {
   subcategoryUrls: boolean
   absoluteLinks: boolean
   searchQuery: any
+  mediaSize: MediaSize
   customDomain: string
   customDomainSlug: string
 }
@@ -129,6 +130,7 @@ WordpressSearchResultBlock.defaultProps = {
   showExcerpts: false,
   subcategoryUrls: false,
   absoluteLinks: false,
+  mediaSize: undefined,
   customDomain: undefined,
   customDomainSlug: undefined,
 }
@@ -222,6 +224,14 @@ const messages = defineMessages({
     defaultMessage: '',
     id: 'admin/editor.wordpressSubcategoryUrls.description',
   },
+  mediaSizeTitle: {
+    defaultMessage: '',
+    id: 'admin/editor.wordpressMediaSize.title',
+  },
+  mediaSizeDescription: {
+    defaultMessage: '',
+    id: 'admin/editor.wordpressMediaSize.description',
+  },
 })
 
 WordpressSearchResultBlock.schema = {
@@ -296,6 +306,15 @@ WordpressSearchResultBlock.schema = {
       title: messages.subcategoryUrlsTitle.id,
       description: messages.subcategoryUrlsDescription.id,
       type: 'boolean',
+      isLayout: false,
+      default: '',
+    },
+    mediaSize: {
+      title: messages.mediaSizeTitle.id,
+      description: messages.mediaSizeDescription.id,
+      type: 'string',
+      enum: ['thumbnail', 'medium', 'medium_large', 'large', 'full'],
+      enumNames: ['Thumbnail', 'Medium', 'Medium Large', 'Large', 'Full'],
       isLayout: false,
       default: '',
     },